refactor(perf): modular api to ts
